+++ speakers/authapp/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
+
 from .models import *
 from .forms import CustomUserCreationForm, CustomUserChangeForm
 
@@ -29,34 +30,4 @@ admin.site.register(UserProfile, UserProfileAdmin)
 
 @admin.register(Token)
 class TokenAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(City)
-class CityAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(Person)
-class PersonAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(Company)
-class CompanyAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(Projects)
-class ProjectsAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(ProjectAdmin)
-class ProjectAdminAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(ProjectLecturer)
-class ProjectLecturerAdmin(admin.ModelAdmin):
     pass
