+++ speakers/workrooms/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views.person import *
+
+urlpatterns = [
+
+]
