+++ speakers/workrooms/views/person.py
@@ -1,0 +1,2 @@
+from rest_framework.views import APIView
+
